Apply simplify pass: shared dialog host harness, named rename response, neutral flag name
