+++ test/components/command-dialog-reattach-reset.test.ts
@@ -4,7 +4,7 @@
  * The command dialog is a reused singleton: clicking a job in firmware-tasks
  * reattaches via the public followJob without going through open()/startCommand.
  * That path must clear per-session flags, else a prior install's missing-upload
- * failure leaves _installMissingUpload set and suppresses the clean/reset hint
+ * failure leaves _compileMissingDependent set and suppresses the clean/reset hint
  * on a subsequently-replayed real build failure. Pin the reset here.
  */
 import { describe, expect, it } from "vitest";
@@ -13,20 +13,20 @@ import { ESPHomeCommandDialog } from "../../src/components/command-dialog.js";
 import { makeFirmwareJob } from "../_make-firmware-job.js";
 
 describe("command-dialog public followJob state reset", () => {
-  it("clears a stale _installMissingUpload on reattach", () => {
+  it("clears a stale _compileMissingDependent on reattach", () => {
     const el = new ESPHomeCommandDialog() as unknown as {
-      _installMissingUpload: boolean;
+      _compileMissingDependent: boolean;
       _streamId: string;
       _api: { firmwareFollowJob: () => string };
       followJob: (job: FirmwareJob, displayName: string) => void;
     };
     // Left over from a prior install whose compile succeeded but found no upload.
-    el._installMissingUpload = true;
+    el._compileMissingDependent = true;
     el._streamId = "";
     el._api = { firmwareFollowJob: () => "stream-1" } as never;
 
     el.followJob(makeFirmwareJob({ job_id: "j1", job_type: JobType.COMPILE }), "device");
 
-    expect(el._installMissingUpload).toBe(false);
+    expect(el._compileMissingDependent).toBe(false);
   });
 });
