+++ test/components/command-dialog-reset-suggestion.test.ts
@@ -42,7 +42,7 @@ interface Host {
   _userStopped: boolean;
   _commandType: string;
   _failedDuringValidate: boolean;
-  _installMissingUpload: boolean;
+  _compileMissingDependent: boolean;
   _jobId: string;
   _jobs: Map<string, FirmwareJob>;
   _primedSource: {
@@ -62,7 +62,7 @@ function baseHost(overrides: Partial<Host> = {}): Host {
     _userStopped: false,
     _commandType: "install",
     _failedDuringValidate: false,
-    _installMissingUpload: false,
+    _compileMissingDependent: false,
     _jobId: "job-1",
     _jobs: new Map(),
     _primedSource: null,
@@ -94,7 +94,7 @@ describe("renderResetSuggestion — local build", () => {
 
   it("renders nothing when the install compile succeeded but had no upload", () => {
     // The compile was fine; clean/reset wouldn't help a missing upload step.
-    const host = baseHost({ _installMissingUpload: true });
+    const host = baseHost({ _compileMissingDependent: true });
     expectNoSuggestion(render(host));
   });
 });
